+++ src/commands/cmd.user.cpp
@@ -2,7 +2,7 @@
 
 bool requestPlayersProfile(DataObject &data, bool &verbose)
 {
-    if (verbose) std::clog << "LOG:" << "[ Start ] [ Thread ] getUserInfo\n";
+    if (verbose) spdlog::info("[ Start ] [ Thread ] getUserInfo");
 
     bool result = false;
 
@@ -13,15 +13,14 @@ bool requestPlayersProfile(DataObject &data, bool &verbose)
             std::this_thread::sleep_for(std::chrono::milliseconds(CSGO_CLI_STEAM_HELLO_DELAY));
 
             CSGOMMHello mmhello;
-            if (verbose) std::clog << "LOG:" << "          Requesting: Hello\n";
+            if (verbose) spdlog::info("          Requesting: Hello");
             mmhello.RefreshWait();
-            if (verbose) std::clog << "LOG:" << "          Got Hello\n";
+            if (verbose) spdlog::info("          Got Hello");
 
             result = true;
 
-            if (verbose) std::clog << "DEBUG:" << mmhello.data.ShortDebugString();
-            if (verbose) std::clog << "DEBUG:" << mmhello.data.DebugString();
-            if (verbose) std::clog << "DEBUG:" << mmhello.data.medals().DebugString();
+            if (verbose) spdlog::debug("mmhello.data.ShortDebugString: {}", mmhello.data.ShortDebugString());
+            if (verbose) spdlog::debug("mmhello.data.DebugString {}", mmhello.data.DebugString());
 
             // player level
             data.player_level       = mmhello.data.player_level();
@@ -62,7 +61,7 @@ bool requestPlayersProfile(DataObject &data, bool &verbose)
             printError("Fatal error", e.what());
             result = false;
         }
-        if (verbose) std::clog << "LOG:" << "[ End   ] [ Thread ] getUserInfo\n";
+        if (verbose) spdlog::info("[ End   ] [ Thread ] getUserInfo");
         return 0;
     });
 
@@ -75,53 +74,46 @@ void printPlayersProfile(DataObject &data)
 {
     // ---------- Format Output Strings
 
-    char name[40];
-    sprintf(name, "%ls", data.playername); // %ls format = wchar_t*
+    std::string rank = fmt::format("{} ({}/18)", data.getPlayerRank(), data.rank_id);
 
-    const auto steam_profile_url = fmt::format("https://steamcommunity.com/profiles/{}", data.steam_id);
+    std::string level = fmt::format("{0} ({1}/40) (XP: {2}/5000 | {3:.2f}%)", data.getPlayerLevel(), data.player_level, data.getPlayerXp(), data.getPlayerXpPercentage());
 
-    const auto rank = fmt::format("{} ({}/18)", data.getPlayerRank(), data.rank_id);
+    std::string likes = fmt::format("{} x friendly, {} x teaching, {} x leader", data.cmd_friendly, data.cmd_teaching, data.cmd_leader);
 
-    const auto level = fmt::format("{0} ({1}/40) (XP: {2}/5000 | {3:.2f}%)", data.getPlayerLevel(), data.player_level, data.getPlayerXp(), data.getPlayerXpPercentage());
+    std::string penalty = fmt::format("{} ({} Minutes)", data.penalty_reason, (data.penalty_seconds / 60));
 
-    const auto likes = fmt::format("{} x friendly, {} x teaching, {} x leader", data.cmd_friendly, data.cmd_teaching, data.cmd_leader);
+    std::string clan = fmt::format("{} \"{}\"", data.clan_name,  data.clan_tag);
 
-    const auto penalty = fmt::format("{} ({} Minutes)", data.penalty_reason, (data.penalty_seconds / 60));
-
-    const auto clan = fmt::format("{} \"{}\"", data.clan_name,  data.clan_tag);
-
-    // TODO how to access medals data? they were available, but are now gone...
+    // TODO how to access medals data?
     //auto medals = fmt::format("{} x arms, {} x combat, {} x global, {} x team, {} x weapon",
     //    data.medals_arms, data.medals_combat, data.medals_global, data.medals_team, data.medals_weapon);
 
     // ---------- Output Table
 
-    fmt::print("\n Hello {}!\n\n Here is your user profile:\n", name);
+    const auto printAligned {[=](const std::string &a, const std::string &b = "") {
+      return fmt::print(" {0:<18} {1}\n", a, b);
+    }};
 
-    tabulate::Table t;
-    t.column(1).format().multi_byte_characters(true); // for name
-    t.column(0).format().font_align(FontAlign::left);
-    t.column(1).format().font_align(FontAlign::center);
+    fmt::print("\n Hello {}!\n", data.playername);
+    fmt::print("\n Here is your user profile:\n\n");
 
-    t.add_row( row_t{" [Steam]"                                                         });
-    t.add_row( row_t{""                                                                 });
-    t.add_row( row_t{" Name:"               , name                                      });
-    t.add_row( row_t{" Clan:"               , clan                                      });
-    t.add_row( row_t{" ID:"                 , toSteamIDClassic(data.steam_id)           });
-    t.add_row( row_t{" ID32:"               , toSteamID32(data.steam_id)                });
-    t.add_row( row_t{" ID64:"               , std::to_string(data.steam_id)             });
-    t.add_row( row_t{" Player Level:"       , std::to_string(data.steam_player_level)   });
-    t.add_row( row_t{" VAC Status:"         , data.getVacStatus()                       });
-    t.add_row( row_t{" Profile URL:"        , steam_profile_url                         });
-    t.add_row( row_t{""                                                                 });
-    t.add_row( row_t{" [CS:GO]"                                                         });
-    t.add_row( row_t{""                                                                 });
-    t.add_row( row_t{" Rank:"               , rank                                      });
-    t.add_row( row_t{" MatchMaking Wins:"   , std::to_string(data.rank_wins)            });
-    t.add_row( row_t{" Player Level:"       , level                                     });
-    t.add_row( row_t{" Likes:"              , likes                                     });
-    t.add_row( row_t{" Penalty:"            , penalty                                   });
-    //t.add_row( row_t{" Medals:"             , medals                                  });
-
-    std::cout << t << std::endl;
+    printAligned("[Steam]"                                                        );
+    printAligned(" "                                                              );
+    printAligned("Name:"                , data.playername                         );
+    printAligned("Clan:"                , clan                                    );
+    printAligned("ID:"                  , toSteamIDClassic(data.steam_id)         );
+    printAligned("ID32:"                , toSteamID32(data.steam_id)              );
+    printAligned("ID64:"                , std::to_string(data.steam_id)           );
+    printAligned("Player Level:"        , std::to_string(data.steam_player_level) );
+    printAligned("VAC Status:"          , data.getVacStatus()                     );
+    printAligned("Profile URL:"         , data.getSteamProfileUrl()               );
+    printAligned(" "                                                              );
+    printAligned("[CS:GO]"                                                        );
+    printAligned(" "                                                              );
+    printAligned("Rank:"                , rank                                    );
+    printAligned("MatchMaking Wins:"    , std::to_string(data.rank_wins)          );
+    printAligned("Player Level:"        , level                                   );
+    printAligned("Likes:"               , likes                                   );
+    printAligned("Penalty:"             , penalty                                 );
+    //printAligned("Medals:"              , medals                                  );
 }