+++ src/csgo/CSGOClient.h
@@ -2,6 +2,7 @@
 #define CSGOClient_H
 
 // Steamworks SDK
+#pragma warning(disable: 4996)
 #include <steam/steam_api.h>
 #include <steam/isteamgamecoordinator.h>