+++ src/csgo/CSGOMatchList.cpp
@@ -31,19 +31,6 @@ CSGOMatchList::~CSGOMatchList()
 void CSGOMatchList::OnMatchList(const CMsgGCCStrike15_v2_MatchList& msg)
 {
     std::unique_lock<std::mutex> lock(m_matchMutex);
-    /*matchList = msg;
-    size_t oldCount = m_matches.size();
-    for (auto it = msg.matches().rbegin(); it != msg.matches().rend(); ++it)
-    {
-        auto oldmatch = std::find_if(m_matches.begin(), m_matches.end(), [it](CDataGCCStrike15_v2_MatchInfo& info)
-        {
-            return info.matchid() == it->matchid();
-        });
-
-        if (oldmatch == m_matches.end()) {
-            m_matches.push_back(*it);
-        }
-    }*/
 
     for (auto it = msg.matches().rbegin(); it != msg.matches().rend(); ++it) {
         m_matches.push_back(*it);
@@ -61,8 +48,9 @@ void CSGOMatchList::Refresh()
     CMsgGCCStrike15_v2_MatchListRequestRecentUserGames request;
     request.set_accountid(accountId);
 
-    if (CSGOClient::GetInstance()->SendGCMessage(k_EMsgGCCStrike15_v2_MatchListRequestRecentUserGames, &request) != k_EGCResultOK)
-        throw ExceptionHandler("Failed to send EMsgGCCStrike15_v2_MatchListRequestRecentUserGames");
+    if (CSGOClient::GetInstance()->SendGCMessage(k_EMsgGCCStrike15_v2_MatchListRequestRecentUserGames, &request) != k_EGCResultOK) {
+      throw ExceptionHandler("Failed to send EMsgGCCStrike15_v2_MatchListRequestRecentUserGames");
+    }
 }
 
 void CSGOMatchList::RefreshWait()
@@ -85,7 +73,7 @@ const std::vector<CDataGCCStrike15_v2_MatchInfo>& CSGOMatchList::Matches() const
 }
 
 
-int CSGOMatchList::getOwnIndex(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats)
+int CSGOMatchList::getOwnIndex(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats) const
 {
     uint32 accountId = SteamUser()->GetSteamID().GetAccountID();
 
@@ -95,35 +83,39 @@ int CSGOMatchList::getOwnIndex(const CMsgGCCStrike15_v2_MatchmakingServerRoundSt
         }
     }
 
-    throw "unable to find own AccountID in matchinfo";
+    throw ExceptionHandler("Unable to find own AccountID in matchinfo.");
 }
 
-int CSGOMatchList::getPlayerIndex(uint32 accid, const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats)
+int CSGOMatchList::getPlayerIndex(uint32 accountId, const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats) const
 {
     for (int i = 0; i < roundStats.reservation().account_ids().size(); ++i) {
-        if (roundStats.reservation().account_ids(i) == accid) {
+        if (roundStats.reservation().account_ids(i) == accountId) {
             return i;
         }
     }
 
-    throw "unable to find specified AccountID in matchinfo";
+    throw ExceptionHandler("Unable to find specified AccountID in matchinfo.");
 }
 
-std::string CSGOMatchList::getMatchResult(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats)
+std::string CSGOMatchList::getMatchResult(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats) const
 {
-    int num = getMatchResultNum(roundStats);
+    const int num = getMatchResultNum(roundStats);
     if (num == 0) return "TIE";
     if (num == 1) return "WIN";
     return "LOSS";
 }
 
-int CSGOMatchList::getMatchResultNum(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats)
+int CSGOMatchList::getMatchResultNum(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats& roundStats) const
 {
-    if (roundStats.match_result() == 0)
-        return 0; // tie
-    if (roundStats.match_result() == 1 && getOwnIndex(roundStats) <= 4)
-        return 1; // win
-    if (roundStats.match_result() == 2 && getOwnIndex(roundStats) >= 5)
-        return 1; // win
+    const auto ownIndex = getOwnIndex(roundStats);
+    if (roundStats.match_result() == 0) {
+      return 0; // tie
+    }
+    if (roundStats.match_result() == 1 && ownIndex <= 4) {
+      return 1; // win
+    }
+    if (roundStats.match_result() == 2 && ownIndex >= 5) {
+      return 1; // win
+    }
     return 2; // loss
 }