+++ src/csgo/CSGOMatchList.h
@@ -20,17 +20,15 @@ public:
 
     const std::vector<CDataGCCStrike15_v2_MatchInfo>& Matches() const;
 
-    int getOwnIndex(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats);
-    int getPlayerIndex(uint32 accid, const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats);
-    std::string getMatchResult(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats);
-    int getMatchResultNum(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats);
+    int getOwnIndex(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats) const;
+    int getPlayerIndex(uint32 accid, const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats) const;
+    std::string getMatchResult(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats) const;
+    int getMatchResultNum(const CMsgGCCStrike15_v2_MatchmakingServerRoundStats & roundStats) const;
 
     CMsgGCCStrike15_v2_MatchList matchList;
 
 private:
     void OnMatchList(const CMsgGCCStrike15_v2_MatchList& msg);
-    //void LoadOldMatchList();
-    //void SaveMatchList();
 
 private:
     bool m_updateComplete = false;